MNT: Move AsyncInput class to utils.
